* configure.in (AM_INIT_AUTOMAKE): Bump version.
[This change should fix bug #8391]
* NEWS: Add note about this bug fix.
* include/avr/boot.h (__BOOT_LOCK_BITS_MASK): Remove.
(__boot_lock_bits_set): Use the compliment of lock_bits arg so the user
specifies only which bits are to be set. NOTE that this changes the
usage but the previous usage was confusing and not well documented.
(__boot_lock_bits_set_alternate): Ditto.
(boot_lock_bits_set): Improve documentation.